# Performance fix for allreduce on large buffers