--- a/gloo/benchmark/main.cc
+++ b/gloo/benchmark/main.cc
@@ -9,6 +9,7 @@
 #include <memory>
 
 #include "gloo/allgather_ring.h"
+#include "gloo/allreduce.h"
 #include "gloo/allreduce_halving_doubling.h"
 #include "gloo/allreduce_bcube.h"
 #include "gloo/allreduce_ring.h"
@@ -195,6 +196,74 @@ class ReduceScatterBenchmark : public Benchmark<T> {
 
 } // namespace
 
+// Namespace for the new style algorithm benchmarks.
+namespace {
+
+template <typename T>
+class NewAllreduceBenchmark : public Benchmark<T> {
+  using allocation = std::vector<std::vector<T, aligned_allocator<T, kBufferAlignment>>>;
+
+ public:
+  NewAllreduceBenchmark(
+    std::shared_ptr<::gloo::Context>& context,
+    struct options& options)
+      : Benchmark<T>(context, options),
+        opts_(context) {}
+
+  allocation newAllocation(int inputs, size_t elements) {
+    allocation out;
+    out.reserve(inputs);
+    for (size_t i = 0; i < inputs; i++) {
+      out.emplace_back(elements);
+    }
+    return out;
+  }
+
+  void initialize(size_t elements) override {
+    inputAllocation_ = newAllocation(this->options_.inputs, elements);
+    outputAllocation_ = newAllocation(this->options_.inputs, elements);
+
+    // Stride between successive values in any input.
+    const auto stride = this->context_->size * this->options_.inputs;
+    for (size_t i = 0; i < this->options_.inputs; i++) {
+      // Different for every input at every node. This means all
+      // values across all inputs and all nodes are different and we
+      // can accurately detect correctness errors.
+      const auto value = (this->context_->rank * this->options_.inputs) + i;
+      for (size_t j = 0; j < elements; j++) {
+        inputAllocation_[i][j] = (j * stride) + value;
+      }
+    }
+
+    // Generate vectors with pointers to populate the options struct.
+    std::vector<T*> inputPointers;
+    std::vector<T*> outputPointers;
+    for (size_t i = 0; i < this->options_.inputs; i++) {
+      inputPointers.push_back(inputAllocation_[i].data());
+      outputPointers.push_back(outputAllocation_[i].data());
+    }
+
+    // Configure AllreduceOptions struct
+    opts_.setInputs(inputPointers, elements);
+    opts_.setOutputs(outputPointers, elements);
+    opts_.setAlgorithm(AllreduceOptions::Algorithm::RING);
+    void (*fn)(void*, const void*, const void*, long unsigned int) = &sum<T>;
+    opts_.setReduceFunction(fn);
+  }
+
+  void run() override {
+    allreduce(opts_);
+  }
+
+ private:
+  AllreduceOptions opts_;
+
+  allocation inputAllocation_;
+  allocation outputAllocation_;
+};
+
+}
+
 #define RUN_BENCHMARK(T)                                                   \
   Runner::BenchmarkFn<T> fn;                                               \
   if (x.benchmark == "allgather_ring") {                                   \
@@ -248,8 +317,33 @@ class ReduceScatterBenchmark : public Benchmark<T> {
   Runner r(x);                                                             \
   r.run(fn);
 
+template <typename T>
+void runNewBenchmark(options& options) {
+  Runner::BenchmarkFn<T> fn;
+
+  const auto name = options.benchmark.substr(4);
+  if (name == "allreduce_ring") {
+    fn = [&](std::shared_ptr<Context>& context) {
+      return gloo::make_unique<NewAllreduceBenchmark<T>>(context, options);
+    };
+  } else {
+    GLOO_ENFORCE(false, "Invalid benchmark name: ", options.benchmark);
+  }
+
+  Runner runner(options);
+  runner.run(fn);
+}
+
 int main(int argc, char** argv) {
   auto x = benchmark::parseOptions(argc, argv);
+
+  // Run new style benchmarks if the benchmark name starts with "new_".
+  // Eventually we'd like to deprecate all the old style ones...
+  if (x.benchmark.substr(0, 4) == "new_") {
+    runNewBenchmark<float>(x);
+    return 0;
+  }
+
   if (x.benchmark == "pairwise_exchange") {
     RUN_BENCHMARK(char);
   } else if (x.halfPrecision) {