@pietern 2019-07-29
@mrshenli Simplified the `segmentBytes` computation and added comments.